Tests/frontend/sad path utub tests (#357)
* Convert MainPageLocators to HomePageLocators

* Sad path tests added for create UTub

* Handle functionality for showing/hiding UTub Desc Create when it is empty

* Sad path tests for update utub name and description

* Sad path tests for deleting a UTub

Adding sad path tests for UTub UI on desktop.

Fixed UTub Tag creation button to not show when no UTub is selected.
Fixed UTub Description creation when the UTub description is empty. It now shows on hover of UTub Title, and hides when exiting the header of the URL deck.